make pouches spawn in modded structures too
here's a small change with the pouch spawning. Now at mod startup, environmental will add vanilla mineshaft and stronghold to a list. Then check if each mod is on and if so, add their stronghold structures/replacements or mineshafts to the list as well. And when mobs are spawning, it will check all structures in the list to see if they are within any of the structure's bounds.

You can even hook up a config to this to let people change what structures can spawn pouch mobs as well! This is technically *slightly* more efficient when it is just environmental on as if a mob is in a mineshaft, it will set the boolean to true right away and not also check the stronghold since there's no need to anymore.

Hope this helps!